Finished PowerUpBase Class
I also added Invincibility Functionality in the Player class. Basically the Invincible PowerUp would be able to toggle the invincible boolean in the player, and while the invincible boolean is activated, the player's current health will always be set back to it's max health. Then, during the powerdown method, the invincible boolean can be toggled off.