Boss_Phase_Two: Part One
This is the first part of implementing the Boss Phase Two mechanics. So far, I got the Enemy Spawning System to shut off and the Boss to move to the the middle of the screen. Now I gotta figure out how to move the spawning location as well as making the boss shoot.